dockcross-script: use tagged version
Use the tagged version of image in the dockcross bash script.

Previously the `latest` version was always used. This is problematic
because 1) a different version of the image can used than the one that
was used to generate the script, and 2) if the script is moved to a
different system without the `latest` version or the `latest` image is
removed, the script will fail to run.

Pass the missing VERSION ARG and fix up a few images that had missing or
incorrect label metadata.